requirements.txt
pyproj
dash
dash-core-components
dash-html-components
dash-leaflet
dash-table
pandas
gunicorn
dash_bootstrap_components
dash_auth
dash_ag_grid
diskcache
reportlab
Pillow
celery
redis
flask
dash[diskcache]
kaleido>=0.2.1
plotly>=5.0.0
flask_login
holidays
dropbox
python-dotenv
psycopg2
dash-daq
gzip